agrego cosas

+++ src/controllers/products.controller.js
@@ -11,37 +11,42 @@ import EErrors from "../utils/errors/Enum.js";
 export const getProducts = async (req, res) => {
     try {
         const { page, query, limit, order } = req.query;
-    let sortBy;
-    if (order === "desc") {
-        sortBy = -1;
-    } else if (order === "asc") {
-        sortBy = 1;
-    }
-    let products;
-    if (!query) {
-        products = await productsModel.paginate(
-            {},
-            {
-                limit: limit ?? 3,
-                lean: true,
-                page: page ?? 1,
-                sort: { price: sortBy },
-            }
-        );
-    } else {
-        products = await productsModel.paginate(
-            { category: query },
-            {
-                limit: limit ?? 3,
-                lean: true,
-                page: page ?? 1,
-                sort: { price: sortBy },
-            }
-        );
-    }
-    res.render("products", { products, query, order, user: req.session.user });
-    } catch(error) {
-        req.logger.error(`Interval server error getting products ${error}`)
+        let sortBy;
+        if (order === "desc") {
+            sortBy = -1;
+        } else if (order === "asc") {
+            sortBy = 1;
+        }
+        let products;
+        if (!query) {
+            products = await productsModel.paginate(
+                {},
+                {
+                    limit: limit ?? 3,
+                    lean: true,
+                    page: page ?? 1,
+                    sort: { price: sortBy },
+                }
+            );
+        } else {
+            products = await productsModel.paginate(
+                { category: query },
+                {
+                    limit: limit ?? 3,
+                    lean: true,
+                    page: page ?? 1,
+                    sort: { price: sortBy },
+                }
+            );
+        }
+        res.render("products", {
+            products,
+            query,
+            order,
+            user: req.session.user,
+        });
+    } catch (error) {
+        req.logger.error(`Interval server error getting products ${error}`);
         res.status(500).send("Error");
     }
 };
@@ -60,7 +65,9 @@ export const getProductById = async (req, res) => {
         }
         res.send(pFound);
     } catch (error) {
-        req.logger.error(`Interval server error getting products by id ${error}`)
+        req.logger.error(
+            `Interval server error getting products by id ${error}`
+        );
         res.status(500).send("Error");
     }
 };
@@ -86,6 +93,7 @@ export const mockingProducts = async (req, res) => {
 };
 
 export const addProduct = async (req, res) => {
+    //todo agregar una vista que permita agregar productos a un admin o premium y asi se agregue su email como owner directamente
     const product = req.body;
     const addedProduct = await productRepository.addProduct(product);
     if (!addedProduct) {
@@ -102,17 +110,19 @@ export const addProduct = async (req, res) => {
 export const updateProduct = async (req, res) => {
     const prodID = req.params.pid;
     const prodToAdd = req.body;
-    const prodToUpdate = await productRepository.updateProduct(
-        prodID,
-        prodToAdd
-    );
-    if(!prodToUpdate) {
+    const user = req.session.user;
+    console.log(user, "user", prodToAdd, "prod");
+    let prodToUpdate;
+    if (user.role === "Admin" || user.email === prodToAdd.owner) {
+        prodToUpdate = await productRepository.updateProduct(prodID, prodToAdd);
+        console.log(prodToUpdate, "prodtoupdate");
+    } else {
         CustomError.createError({
             name: "Request error",
             cause: addProductError(),
             code: EErrors.ROUTING_ERROR,
             message: "Could not update product",
-        })
+        });
     }
     res.send(prodToUpdate);
 };
@@ -120,18 +130,24 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     try {
         const prodID = req.params.pid;
-        const productToDelete = await productRepository.deleteProduct(prodID);
-        if(!productToDelete) {
+        // console.log("PRODUCTO ID", prodID);
+        const user = req.session.user;
+        // console.log("USUARIO", user);
+        const prodToDelete = await productRepository.getProductById(prodID);
+        // console.log("PRODUCTO TO DELETE", prodToDelete);
+        if (user.role === "Admin" || user.email === prodToDelete.owner) {
+            await productRepository.deleteProduct(prodID);
+        } else {
             CustomError.createError({
                 name: "Request error",
                 cause: productError(),
                 code: EErrors.ROUTING_ERROR,
                 message: "Could not delete product",
-            })
+            });
         }
-        res.send(productToDelete);
+        res.send(prodToDelete);
     } catch (error) {
-        req.logger.error(`Interval server error deleting products ${error}`)
+        req.logger.error(`Interval server error deleting products ${error}`);
         res.status(500).send("Error getting data.");
     }
 };
